introduce Signature method to Vector
Introduce `Signature() interface{}` method to Vector
to help with equality and map lookups. This method
helps resolve the trouble before of figuring out
how to compare Vectors and for that matter when
centroids are no longer changing and clusters are stable.

Also introduce a small app that will used for map
visualizations using mapbox's JS API. When free, I'll
write a small web server to serve back the content
for the visualizations.

Signed-off-by: Emmanuel Odeke <emm.odeke@gmail.com>